+++ reduct/http.py
@@ -14,6 +14,8 @@ API_PREFIX = "/api/v1"
 class HttpClient:
     """Wrapper for HTTP calls"""
 
+    FILE_SIZE_FOR_100_CONTINUE = 256_000
+
     def __init__(
         self, url: str, api_token: Optional[str] = None, timeout: Optional[float] = None
     ):
@@ -23,6 +25,7 @@ class HttpClient:
             {"Authorization": f"Bearer {api_token}"} if api_token is not None else {}
         )
         self.timeout = ClientTimeout(timeout)
+        self.request_count = 0
 
     @asynccontextmanager
     async def request(
@@ -35,7 +38,7 @@ class HttpClient:
         if "content_length" in kwargs:
             content_length = kwargs["content_length"]
             extra_headers["Content-Length"] = str(content_length)
-            if content_length > 256_000:
+            if content_length > self.FILE_SIZE_FOR_100_CONTINUE:
                 # Use 100-continue for large files
                 extra_headers["Expect"] = "100-continue"
 
@@ -51,7 +54,10 @@ class HttpClient:
                     extra_headers[f"x-reduct-label-{name}"] = str(value)
             del kwargs["labels"]
 
-        async with aiohttp.ClientSession(timeout=self.timeout) as session:
+        connector = aiohttp.TCPConnector(force_close=True)
+        async with aiohttp.ClientSession(
+            timeout=self.timeout, connector=connector
+        ) as session:
             try:
                 async with session.request(
                     method,